gem jewelry refactor, among other things

+++ src/main/java/net/solunareclipse1/magitekkit/client/render/LayerHalo.java
@@ -45,12 +45,11 @@ public class LayerHalo extends RenderLayer<AbstractClientPlayer, PlayerModel<Abs
 			if (!(stack.getItem() instanceof GemJewelryBase) || stack.isDamaged()) return;
 			if (stack.getItem() instanceof GemAmulet) {
 				GemAmulet amulet = (GemAmulet) stack.getItem();
-				emcLevel = (float) amulet.getStoredEmc(stack) / amulet.getMaximumEmc(stack);
+				emcLevel = (float)amulet.getStoredEmc(stack) / (float)amulet.getMaximumEmc(stack);
 			}
 		}
 		if (emcLevel <= 0) return;
 		int timer = Math.round(ageInTicks);
-		//if (timer % 20 == 0) System.out.println("EMC Percentage: "+emcLevel);
 		poseStack.pushPose();
 		renderer.getModel().jacket.translateAndRotate(poseStack);
 		poseStack.mulPose(Vector3f.XP.rotationDegrees(270)); // rotate upright
